fix: move denom convetion to dedicated util
